Added a wait call of 3 seconds in SPA tests (#4)
This is to fix an issue testing the plain JavaScript samples.

For some reason, Puppeteer finds the #profileDropDown element on the
page even though it is invisible, but then fails when it tries to
click on it. Waiting for a few seconds gets around this, as by that time
the element should be truly visible on the page.